[fix/#123] 반투명 스낵바 -> DDanSnackBar로 수정
